+++ packages/core/client/src/schema-component/antd/association-filter/hook.ts
@@ -14,7 +14,7 @@ import { useBlockHeightProps } from '../../../block-provider';
 export const useAssociationFilterHeight = () => {
   const height = useDataBlockHeight();
   const { token } = theme.useToken();
-  const { heightProps } = useBlockHeightProps() || {};
+  const { heightProps } = useBlockHeightProps?.() || {};
   const { title } = heightProps || {};
   const blockTitleHeaderHeight = title ? token.fontSizeLG * token.lineHeightLG + token.padding * 2 - 1 : 0;
   return height - 2 * token.paddingLG - blockTitleHeaderHeight;
